% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_window_day.R
\name{add_window_day}
\alias{add_window_day}
\title{Add Window Day Variable}
\usage{
add_window_day(data, window_day_col, window)
}
\arguments{
\item{data}{A data frame containing patient data with columns \code{unique_pt_id}, \code{seqnum}, \code{day}, and \code{bcx_daily}.}

\item{window_day_col}{A string specifying the name of the new column to be added.}

\item{window}{An integer specifying the number of days before and after each day to check for blood culture days.}
}
\value{
A data frame with the new column indicating the presence of blood culture days within the specified window.
}
\description{
This function adds a new column to the data indicating whether there is any blood culture day within a specified window around each day.
}
\examples{
data <- data.frame(
  unique_pt_id = c(1, 1, 1, 2, 2, 2),
  seqnum = c(12602, 12602, 12602, 18613, 18613, 18613),
  day = c(1, 2, 3, 1, 2, 3),
  bcx_daily = c(0, 1, 0, 0, 1, 0)
)
add_window_day(data, "window_day", 1)
}
