% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_new_abx.R
\name{define_new_abx}
\alias{define_new_abx}
\title{Define New Antimicrobial Definition}
\usage{
define_new_abx(daily_data)
}
\arguments{
\item{daily_data}{A data frame containing daily patient data with columns \code{unique_pt_id}, \code{seqnum}, \code{day}, \code{death}, \code{ALL_DAYS}, \code{new_abx_start} and other clinical variables.}
}
\description{
This function creates the variable \code{new_abx_start} to indicate the start of new antimicrobial administration.
}
\examples{
# Example daily data frame
daily_data <- data.frame(
  unique_pt_id = c(1,1,1, 2,2, 3,3,3, 4,4,4,4,4),
  seqnum       = c(1111,1111,1111, 2222,2222, 3333,3333,3333, 4444,4444,4444,4444,4444),
  day          = c(0,1,2, 0,1, 0,1,2, 0,1,2,3,4),
  death        = c(0,0,0, 0,1, 0,0,0, 0,0,0,0,0),
  ALL_DAYS     = c(3,3,3, 2,2, 3,3,3, 5,5,5,5,5),
  bcx_daily    = c(1,0,0, 0,0, 1,0,0, 0,1,0,0,0),
  vasop_daily  = c(0,1,0, 0,0, 0,1,0, 0,0,1,0,0),
  imv_daily    = c(0,0,1, 0,1, 0,0,1, 0,0,1,0,0),
  lact_daily_hi = c(1.5,2.5,1.8, 1.9,2.1, 1.5,2.5,1.8, 1.5,1.9,2.1,2.1,3.0),
  tbili_daily_hi = c(30,35,40, 32,36, 30,35,40, 30,32,35,36,50),
  tbili_daily_lo = c(30,35,40, 32,36, 30,35,40, 30,32,35,36,50),
  tbili_baseline = c(20,20,20, 25,25, 20,20,20, 25,25,25,25,25),
  creat_daily_hi = c(40,45,50, 35,60, 40,45,50, 35,35,45,60,55),
  creat_daily_lo = c(40,45,50, 35,60, 40,45,50, 35,35,45,60,55),
  creat_baseline = c(20,20,20, 25,25, 20,20,20, 25,25,25,25,25),
  plt_daily_hi   = c(150,80,90, 110,70, 150,80,90, 150,140,70,60,50),
  plt_daily_lo   = c(150,80,90, 110,70, 150,80,90, 150,140,70,60,50),
  plt_baseline   = c(200,200,200, 180,180, 200,200,200, 180,180,180,180,180),
  esrd_icd       = c(0,0,0, 0,0, 0,0,0, 0,0,0,0,0),
  new_abx_start  = c(0,1,0, 0,1, 0,1,0, 0,1,0,0,0),
  abx_daily      = c(0,1,1, 0,1, 0,1,1, 0,1,1,1,1)
)
define_new_abx(daily_data)
}
