% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_by_seqnum.R
\name{search_by_seqnum}
\alias{search_by_seqnum}
\title{Search for Data by Sequence Number}
\usage{
search_by_seqnum(sliced_data_list, target_seqnum)
}
\arguments{
\item{sliced_data_list}{A list of lists, where each list contains \code{unique_pt_id}, \code{seqnum}, and a data frame (\code{data}) with sliced data.}

\item{target_seqnum}{An integer specifying the sequence number to search for.}
}
\value{
A list of data frames that match the specified sequence number. If no matches are found, an empty list is returned.
}
\description{
This function searches through a list of sliced data and returns the data slices that match a given sequence number.
}
\examples{
# Example sliced_data_list
sliced_data_list <- list(
  list(unique_pt_id = 1, seqnum = 12602, data = data.frame(day = 1:3, bcx_daily = c(0, 1, 0))),
  list(unique_pt_id = 2, seqnum = 18613, data = data.frame(day = 1:3, bcx_daily = c(0, 1, 0)))
)
search_by_seqnum(sliced_data_list, 12602)
}
