% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_aod_daily.R
\name{add_aod_daily}
\alias{add_aod_daily}
\title{Add Acute Organ Dysfunction (AOD) Daily Variables}
\usage{
add_aod_daily(
  data,
  window_day_col,
  creat_hi_lo_ratio = 2,
  creat_hi_cutoff = 44,
  tbili_hi_cutoff = 34.2,
  tbili_hi_lo_ratio = 2,
  lact_hi_cutoff = 2,
  plt_lo_cutoff = 100,
  plt_lo_hi_ratio = 0.5
)
}
\arguments{
\item{data}{A data frame containing patient data with columns \code{unique_pt_id}, \code{seqnum}, \code{day}, \code{vasop_daily}, \code{imv_daily}, \code{lact_daily_hi}, \code{tbili_daily_hi}, \code{tbili_baseline}, \code{creat_daily_hi}, \code{creat_baseline}, \code{plt_daily_lo}, \code{plt_baseline}, and \code{esrd_icd}.}

\item{window_day_col}{A string specifying the name of the column indicating the presence of blood culture days within a specified window.}

\item{creat_hi_lo_ratio}{The ratio of high to low creatinine levels to define renal dysfunction (default is 2, per the ASE toolkit).}

\item{creat_hi_cutoff}{The cutoff value for high creatinine levels (default is 44 µmol/L, per the Risk, Injury, Failure, Loss, and End-stage (RIFLE) renal disease classification system guidelines).}

\item{tbili_hi_cutoff}{The cutoff value for high bilirubin levels (default is 34.2 µmol/L, per the ASE toolkit).}

\item{tbili_hi_lo_ratio}{The ratio of high to low bilirubin levels to define liver dysfunction (default is 2, per the ASE toolkit).}

\item{lact_hi_cutoff}{The cutoff value for high lactate levels (default is 2 mmol/L, per the ASE toolkit).}

\item{plt_lo_cutoff}{The cutoff value for low platelet counts (default is 100 10^9/L, per the ASE toolkit).}

\item{plt_lo_hi_ratio}{The ratio of low to high platelet counts to define hematologic dysfunction (default is 0.5, per the ASE toolkit).}
}
\value{
A data frame with new columns for each type of AOD and additional calculations based on the specified window.
}
\description{
This function calculates and adds daily indicators for various types of acute organ dysfunction (AOD) to the data,
based on given criteria for cardiovascular, respiratory, lactate, liver, renal, and hematologic dysfunctions.
}
\examples{
# Example data frame
data <- data.frame(
  unique_pt_id = c(1, 1, 1, 2, 2, 2),
  seqnum = c(12602, 12602, 12602, 18613, 18613, 18613),
  day = c(1, 2, 3, 1, 2, 3),
  vasop_daily = c(0, 1, 0, 0, 1, 0),
  imv_daily = c(0, 0, 1, 1, 0, 1),
  lact_daily_hi = c(1.5, 2.5, 1.8, 1.9, 2.1, 3.0),
  tbili_daily_hi = c(30, 35, 40, 32, 36, 38),
  tbili_baseline = c(20, 20, 20, 25, 25, 25),
  creat_daily_hi = c(40, 45, 50, 35, 60, 55),
  creat_baseline = c(20, 20, 20, 25, 25, 25),
  plt_daily_lo = c(150, 80, 90, 110, 70, 50),
  plt_baseline = c(200, 200, 200, 180, 180, 180),
  esrd_icd = c(0, 0, 0, 0, 0, 0)
)
data <- add_window_day(data, "window_day", 1)
add_aod_daily(data, "window_day")
}
