% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_ed_days.R
\name{process_ed_days}
\alias{process_ed_days}
\title{Preprocess ED Data}
\usage{
process_ed_days(daily_data, ed_inclusion = 1)
}
\arguments{
\item{daily_data}{A data frame containing daily patient data with columns \code{unique_pt_id}, \code{seqnum}, \code{day}, \code{death}, \code{ALL_DAYS}, and other clinical variables.}

\item{ed_inclusion}{An integer (1–4) specifying how to preprocess the ED rows in daily_data:
\describe{
\item{1}{ASE Toolkit default: keep 2 days in ED (i.e., rows with \code{day > -2}.)}
\item{2}{Include all ED rows (no filtering).}
\item{3}{Drop all ED rows (i.e., \code{day < 1}).}
\item{4}{Drop ED rows, but if any ED blood culture is detected,
then assign \code{bcx_daily = 1} on hospital day 1.}
}}
}
\value{
A modified \code{daily_data} data frame with ED rows included or excluded
}
\description{
This function processes daily patient data to include or exclude Emergency Department (ED) rows based on different inclusion rules.
}
\examples{
data_path <- system.file(
  "extdata", "example_input_data", "daily_data_mock.csv",
  package = "identifyASE"
)
daily_data <- read.csv(data_path, stringsAsFactors = FALSE)
process_ed_days(daily_data, ed_inclusion = 1)
}
